This was a tower defense that I made for my java class.
There seems to be some errors now in it since I originally made it. In
order to play you might have to create a map yourself by directly
loading the mapmaker. The roads have to go from the left side of the map
to the right for it to work.